--- src/server/api/controllers/channels.controller.js
@@ -1,11 +1,0 @@
-const knex = require('../../config/db');
-
-const deleteChannelById = async (channelId) => {
-  return knex('channels')
-    .where({ id: channelId })
-    .del();
-};
-
-module.exports = {
-  deleteChannelById,
-};
